Move StatusController to Services and update references
